requirements.txt
# utilities
python-dotenv

# frontend
streamlit

# data
pandas
